+++ src/ingest/auth/reset/post.js
@@ -7,39 +7,39 @@
 
 // load eventhub utils
 const firebase = require('../../../utils/firebase')
+const logger = require('../../../utils/logger')
 const response = require('../../../utils/response')
+
+const source = 'ingest/auth/reset'
 
 module.exports = async (req, res) => {
 	try {
 		// try to reset email (may fail if not found)
 		try {
 			await firebase.sendPasswordResetEmail(req.body.email)
-		} catch (err) {
-			console.warn(
-				'ingest/auth/reset',
-				'failed resetting password',
-				JSON.stringify({
-					email: req.body.email,
-					err,
-				})
-			)
-			return response.badRequest(req, res, {
-				status: err.error && err.error.code ? err.error.code : 500,
-				data: err,
+		} catch (error) {
+			logger.log({
+				level: 'notice',
+				message: 'failed resetting password',
+				source,
+				error,
+				data: { email: req.body.email },
 			})
+
+			return response.badRequest(req, res, { status: error?.error?.code || 500, data: error })
 		}
 
 		// return ok
 		return response.ok(req, res, { valid: true })
-	} catch (err) {
-		console.error(
-			'ingest/auth/reset',
-			'failed to reset password',
-			JSON.stringify({
-				headers: req.headers,
-				error: err.stack || err,
-			})
-		)
-		return response.internalServerError(req, res, err)
+	} catch (error) {
+		logger.log({
+			level: 'error',
+			message: 'failed to reset password',
+			source,
+			error,
+			data: { body: req.body, headers: req.headers },
+		})
+
+		return response.internalServerError(req, res, error)
 	}
 }