+++ src/ingest/subscriptions/get.js
@@ -10,6 +10,8 @@ const logger = require('../../utils/logger')
 const pubsub = require('../../utils/pubsub')
 const response = require('../../utils/response')
 
+const source = 'ingest/subscriptions/get'
+
 module.exports = async (req, res) => {
 	try {
 		// preset vars
@@ -20,7 +22,7 @@ module.exports = async (req, res) => {
 		try {
 			subscription = await pubsub.getSubscription(subscriptionName)
 			subscription = subscription.limited
-		} catch (err) {
+		} catch (error) {
 			return response.notFound(req, res, {
 				status: 404,
 				message: `Subscription '${subscriptionName}' not found`,
@@ -45,7 +47,7 @@ module.exports = async (req, res) => {
 		logger.log({
 			level: 'error',
 			message: 'failed to get subscription',
-			source: 'ingest/subscriptions/get',
+			source,
 			error,
 			data: { params: req.params },
 		})